Simplified new(); Added a case for EPSILON
Using \e for EPSILON. All other escapes are just interpreted
as themselves (e.g., "\n" is just the same as "n", not a newline).

Collapsed all the new_X functions into a single very simple new().
The tradeoff is that now there is no arity checking between the
operator and the arguments list.