Translation to program with multiple entry points
Each input regex is parsed into a tree, then the tree is
translated into a sequence of instructions which is appended
to the overall program.

Match instructions now carry an indicator of which rule matched.
That is, there is a separate Match instruction for each rule.
Currently they just point at their own line number, but will
probably end up holding a rule number that can be used to index
into a collection of actions to perform.

The interpreter has not yet been updated.