Case insensitive matching for char classes
In the end the whole approach to case insensitive matching had
to be redone. Basically, using a control flag on the interpreter
meant that if one branch set a flag, all branches would have to
obey it.

So now case insensitivity is handled at the level of individual
instructions. Since each regex is parsed independently, that
prevents bleeding.

There will be more headaches later on. Once we get into general
Unicode territory, it may no longer be the case that downcasing
the bounds of an interval will yield exactly the set of
characters in the original interval, downcased.