Preliminary attempt to relax input requirements
We are requiring strings, but if whole documents are being
transferred over the net, we might want to be able to start
working before a whole document has arrived. So don't insist
that we have all the characters right up front.

This checkin is just some preliminary spade work, to try to
get things so we are only ever looking at the first character
in a string slice.