Two ways char_at can return None; check them
char_at() can return None if there is a UTF-8 decoding error,
and also if we run out of string (there really is nothing there,
not just a bad character).

We need to know which version of "None" we are looking at. If
we are just out of string, then continue running the program,
since epsilon transitions will still be valid.